Frontend/old/requirements.txt
Pillow==2.7.0
Flask==0.10.1